Experimental ChannelBuffer to stream from Channels
without wrapping them as InputStreams.

While I'm in here, I refactored the common behavior (almost everything,
as it turns out) from BoundedInputStreamBuffer into StreamingBuffer.
